Intégration animation ennemis+ personnage ensemble

+++ Assets/Scripts/Gameplay/HealthManagement/HealthBarController.cs
@@ -7,13 +7,13 @@ public class HealthBarController : MonoBehaviour
 {
     [SerializeField] private Slider slider;
 
-    public void SetMaxHealth(int health)
+    public void SetMaxHealth(float health)
     {
         slider.maxValue = health;
         slider.value = health; ;
     }
 
-    public void SetHealth(int health)
+    public void SetHealth(float health)
     {
         slider.value = health;
     }
